// packages/html/src/types/core.types.ts
export type Key = string | number | null;
export type VNode = VElement | string | number | boolean | null | undefined;
export type Ref<T> = RefObject<T> | ((instance: T | null) => void);

export abstract class Component<P = {}, S = {}> {
    // Add isReactComponent flag for type checking
    static readonly isReactComponent = true;
    props: Readonly<P>;
    state: Readonly<S>;

    constructor(props: P) {
        this.props = Object.freeze({ ...props });
        this.state = Object.freeze({} as S);
    }

    abstract render(): JSX.Element;

    abstract componentDidMount?(): void;

    abstract componentDidUpdate?(prevProps: Readonly<P>, prevState: Readonly<S>): void;

    abstract componentWillUnmount?(): void;

    abstract shouldComponentUpdate?(nextProps: Readonly<P>, nextState: Readonly<S>): boolean;

    abstract setState(
        updater: Partial<S> | ((prevState: Readonly<S>, props: Readonly<P>) => Partial<S>),
        callback?: () => void,
    ): void;
}

export type VNodeType = string | FunctionComponent | typeof Component;

/**
 * Props and elements
 */
export interface Props {
    children?: VNode | VNode[];
    key?: Key;
    ref?: RefObject<any>;
    className?: string;
    style?: Partial<CSSStyleDeclaration>;

    [key: string]: any;
}

export interface VElement {
    type: VNodeType;
    props: Props;
    key?: Key;
    ref?: RefObject<any> | null;
    __owner?: Component | null;
    __context?: Context<any>;
}

/**
 * Function component definition
 */
export interface FunctionComponent<P = {}> {
    defaultProps?: Partial<P>;
    displayName?: string;
    _owner?: Component | null;
    _hooks?: any[];
    _hookIndex?: number;
    _signals?: Set<ISignal<any>> | null;
    setState?: (state: any) => void;

    (props: P): JSX.Element;
}

export type FC<P = {}> = FunctionComponent<P>;

/**
 * Ref types
 */
export interface RefObject<T = any> {
    current: T | null;
}

export interface ForwardRefRenderFunction<T, P = {}> {
    displayName?: string;

    (props: P, ref: RefObject<T>): JSX.Element;
}

export interface ForwardRefExoticComponent<P> extends FunctionComponent<P> {
    defaultProps?: Partial<P>;
    displayName?: string;
}

/**
 * Context types
 */
export interface Context<T> {
    Provider: ProviderComponent<T>;
    Consumer: ConsumerComponent<T>;
    displayName?: string;
    _currentValue: T;
    _defaultValue: T;
    _currentProvider: Component | null;
}
// ...
export interface ISignal<T> {
    value: T;
    subscribers: Set<() => void>;

    subscribe(callback: () => void): () => void;

    unsubscribe(callback: () => void): void;

    emit(value: T): void;
}